requirements.txt
fastapi
uvicorn
python-multipart
yt-dlp
openai-whisper
pydub
requests
python-dotenv
ffmpeg-python